refactor(sync): extract fs-utils and scaffold-merge modules (Steps 1-2)
Modularization of synchronize.mjs — Phase 1 (zero-risk utility splits).

Step 1 — fs-utils.mjs:
  Extracts runConcurrent, ensureDir, writeOutput, walkDir into a stateless
  I/O helpers module. Re-exported from synchronize.mjs for backward compat.

Step 2 — scaffold-merge.mjs:
  Extracts threeWayMerge and normalizeForComparison into a dedicated module.
  Isolates the Prettier-alignment guard so it can be tested independently.

synchronize.mjs shrinks by ~140 lines. No functional changes.
No test import updates required for Steps 1-2.
