+++ .agentkit/engines/node/src/fs-utils.mjs
@@ -1,0 +1,46 @@
+/**
+ * Retort — Filesystem Utilities
+ * Stateless async I/O helpers used by the sync engine and other modules.
+ * No domain knowledge — pure Node.js primitives.
+ */
+import { existsSync } from 'fs';
+import { mkdir, readdir, writeFile } from 'fs/promises';
+import { dirname, join } from 'path';
+
+export async function runConcurrent(items, fn, concurrency = 50) {
+  const chunks = [];
+  for (let i = 0; i < items.length; i += concurrency) {
+    chunks.push(items.slice(i, i + concurrency));
+  }
+  for (const chunk of chunks) {
+    await Promise.all(chunk.map(fn));
+  }
+}
+
+export async function ensureDir(dirPath) {
+  await mkdir(dirPath, { recursive: true });
+}
+
+export async function writeOutput(filePath, content) {
+  await ensureDir(dirname(filePath));
+  await writeFile(filePath, content, 'utf-8');
+}
+
+export async function* walkDir(dir) {
+  if (!existsSync(dir)) return;
+  let entries = [];
+  try {
+    entries = await readdir(dir, { withFileTypes: true });
+  } catch (err) {
+    if (err?.code === 'ENOENT') return;
+    throw err;
+  }
+  for (const entry of entries) {
+    const full = join(dir, entry.name);
+    if (entry.isDirectory()) {
+      yield* walkDir(full);
+    } else {
+      yield full;
+    }
+  }
+}
