+++ .agentkit/engines/node/src/synchronize.mjs
@@ -26,6 +26,8 @@ import {
   loadFeatureSpec,
   resolveFeatures,
 } from './feature-manager.mjs';
+import { ensureDir, runConcurrent, walkDir, writeOutput } from './fs-utils.mjs';
+import { normalizeForComparison, threeWayMerge } from './scaffold-merge.mjs';
 import {
   categorizeFile,
   computeProjectCompleteness,
@@ -80,115 +82,7 @@ export function resolveCommandPath(cmdName, prefix, strategy = 'filename') {
   return { dir: '', stem: `${prefix}-${cmdName}` };
 }
 
-// ---------------------------------------------------------------------------
-// Three-way merge for managed scaffold files
-// ---------------------------------------------------------------------------
-
-/**
- * Performs a three-way merge using git merge-file.
- * @param {string} oursContent - User's current version (disk)
- * @param {string} baseContent - Last generated version (scaffold cache)
- * @param {string} theirsContent - Newly generated version (template)
- * @returns {{ merged: string, hasConflicts: boolean }|null} null if git unavailable
- */
-function threeWayMerge(oursContent, baseContent, theirsContent) {
-  const prefix = join(
-    tmpdir(),
-    `agentkit-merge-${Date.now()}-${Math.random().toString(36).slice(2, 8)}`
-  );
-  const oursFile = `${prefix}-ours`;
-  const baseFile = `${prefix}-base`;
-  const theirsFile = `${prefix}-theirs`;
-
-  writeFileSync(oursFile, oursContent);
-  writeFileSync(baseFile, baseContent);
-  writeFileSync(theirsFile, theirsContent);
-
-  try {
-    const merged = execFileSync(
-      'git',
-      [
-        'merge-file',
-        '-p',
-        '--diff3',
-        '-L',
-        'YOUR_EDITS',
-        '-L',
-        'LAST_SYNC',
-        '-L',
-        'NEW_TEMPLATE',
-        oursFile,
-        baseFile,
-        theirsFile,
-      ],
-      { encoding: 'utf-8' }
-    );
-    return { merged, hasConflicts: false };
-  } catch (err) {
-    if (err.status === 1) {
-      // Merge completed but has conflicts
-      return {
-        merged: typeof err.stdout === 'string' ? err.stdout : oursContent,
-        hasConflicts: true,
-      };
-    }
-    // git merge-file not available or other error
-    return null;
-  } finally {
-    try {
-      unlinkSync(oursFile);
-    } catch {
-      /* ignore */
-    }
-    try {
-      unlinkSync(baseFile);
-    } catch {
-      /* ignore */
-    }
-    try {
-      unlinkSync(theirsFile);
-    } catch {
-      /* ignore */
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Normalize content for semantic comparison (ignores table-cell padding)
-// ---------------------------------------------------------------------------
-
-/**
- * Strips trailing whitespace and normalises markdown table-cell padding so
- * that a Prettier-aligned table and a compact table compare as equal when
- * the cell *values* are identical.  Used to detect whether a disk file
- * differs from the scaffold cache for reasons other than whitespace.
- *
- * @param {string} content
- * @returns {string}
- */
-export function normalizeForComparison(content) {
-  return content
-    .split('\n')
-    .map((line) => {
-      if (/^\s*\|/.test(line)) {
-        // Separator rows (|---|---| or | --- | --- |) — collapse to |---| canonical form
-        if (/^\s*\|[\s|:-]+\|\s*$/.test(line)) {
-          const cols = line.split('|').filter((_, i, a) => i > 0 && i < a.length - 1);
-          return '|' + cols.map((c) => c.trim().replace(/^(:?)-+(:?)$/, '$1-$2')).join('|') + '|';
-        }
-        // Data rows — normalise cell padding to a single space either side
-        return line
-          .split('|')
-          .map((cell, i, arr) =>
-            i === 0 || i === arr.length - 1 ? cell.trimEnd() : ` ${cell.trim()} `
-          )
-          .join('|');
-      }
-      return line.trimEnd();
-    })
-    .join('\n')
-    .trimEnd();
-}
+// threeWayMerge and normalizeForComparison live in scaffold-merge.mjs (imported above)
 
 // ---------------------------------------------------------------------------
 // I/O helpers
@@ -280,43 +174,9 @@ async function readTemplateText(filePath) {
   return content;
 }
 
-export async function runConcurrent(items, fn, concurrency = 50) {
-  const chunks = [];
-  for (let i = 0; i < items.length; i += concurrency) {
-    chunks.push(items.slice(i, i + concurrency));
-  }
-  for (const chunk of chunks) {
-    await Promise.all(chunk.map(fn));
-  }
-}
-
-export async function ensureDir(dirPath) {
-  await mkdir(dirPath, { recursive: true });
-}
-
-export async function writeOutput(filePath, content) {
-  await ensureDir(dirname(filePath));
-  await writeFile(filePath, content, 'utf-8');
-}
-
-export async function* walkDir(dir) {
-  if (!existsSync(dir)) return;
-  let entries = [];
-  try {
-    entries = await readdir(dir, { withFileTypes: true });
-  } catch (err) {
-    if (err?.code === 'ENOENT') return;
-    throw err;
-  }
-  for (const entry of entries) {
-    const full = join(dir, entry.name);
-    if (entry.isDirectory()) {
-      yield* walkDir(full);
-    } else {
-      yield full;
-    }
-  }
-}
+// runConcurrent, ensureDir, writeOutput, walkDir live in fs-utils.mjs (imported above)
+// Re-export for any external callers that imported from synchronize.mjs directly.
+export { ensureDir, runConcurrent, walkDir, writeOutput };
 
 function inferOverlayFromProjectRoot(agentkitRoot, projectRoot) {
   const inferredName = basename(resolve(projectRoot));
